requirements.txt
google-generativeai
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
meteomatics
googlemaps
python-dotenv
flask
APScheduler
winotify
